[update] refs #50 ルーティング変更
 ディレクトリ構成改善のため
