+++ app/(routes)/dashboard/layout.tsx
@@ -1,10 +1,10 @@
 import { ReactNode } from 'react'
 import { createServerComponentClient } from '@supabase/auth-helpers-nextjs'
-import { Database } from '../../lib/database.types'
+import { Database } from '../../../lib/database.types'
 import { cookies } from 'next/headers'
-import Navigation from '../_components/navigation'
-import Content from '../_components/content'
-import SupabaseListener from '../_components/supabase-listener'
+import Navigation from '../../_components/navigation'
+import Content from '../../_components/content'
+import SupabaseListener from '../../_components/supabase-listener'
 
 export default async function Dashboardlayout({
   children,
