+++ app/_components/logout.tsx
@@ -3,7 +3,7 @@
 import { FormEvent, useState } from 'react'
 import { createClientComponentClient } from '@supabase/auth-helpers-nextjs'
 import { useRouter } from 'next/navigation'
-import Loading from '../loading'
+import Loading from '../(routes)/loading'
 import type { Database } from '../../lib/database.types'
 import { Card, CardBody, Button } from './common'
