+++ app/_components/matrix/assets/goal-edit-form.tsx
@@ -10,7 +10,7 @@ import {
 } from '../../common'
 import { Dispatch, SetStateAction, useState } from 'react'
 import { z } from 'zod'
-import Loading from '../../../loading'
+import Loading from '../../../(routes)/loading'
 import { createClientComponentClient } from '@supabase/auth-helpers-nextjs'
 import { Database } from '../../../../lib/database.types'
 import { SubmitHandler, useForm } from 'react-hook-form'
@@ -52,7 +52,7 @@ export default function GoalEditForm({
   setOpen: Dispatch<SetStateAction<boolean>>
   goal: Goal | null
   userId: string | undefined
-  }) {
+}) {
   const router = useRouter()
   const supabase = createClientComponentClient<Database>()
   const [loading, setLoading] = useState(false)
